+++ examples/basic/src/model-function/generate-text-example.ts
@@ -1,11 +1,11 @@
-import { OpenAITextGenerationModel, generateText } from "modelfusion";
+import { OpenAICompletionModel, generateText } from "modelfusion";
 import dotenv from "dotenv";
 
 dotenv.config();
 
 async function main() {
   const text = await generateText(
-    new OpenAITextGenerationModel({
+    new OpenAICompletionModel({
       model: "gpt-3.5-turbo-instruct",
       temperature: 0.7,
       maxCompletionTokens: 500,