+++ examples/basic/src/model-provider/openai/openai-text-generation-api-configuration-example.ts
@@ -1,6 +1,6 @@
 import {
   OpenAIApiConfiguration,
-  OpenAITextGenerationModel,
+  OpenAICompletionModel,
   generateText,
   retryWithExponentialBackoff,
   throttleUnlimitedConcurrency,
@@ -11,7 +11,7 @@ dotenv.config();
 
 async function main() {
   const text = await generateText(
-    new OpenAITextGenerationModel({
+    new OpenAICompletionModel({
       model: "gpt-3.5-turbo-instruct",
       api: new OpenAIApiConfiguration({
         // all parameters are optional: