+++ examples/basic/src/util/api-configuration/retry-exponential-backoff-example.ts
@@ -1,6 +1,6 @@
 import {
   OpenAIApiConfiguration,
-  OpenAITextGenerationModel,
+  OpenAICompletionModel,
   generateText,
   retryWithExponentialBackoff,
 } from "modelfusion";
@@ -19,7 +19,7 @@ async function main() {
   });
 
   const text = await generateText(
-    new OpenAITextGenerationModel({
+    new OpenAICompletionModel({
       model: "gpt-3.5-turbo-instruct",
       api,
     }),