+++ examples/basic/src/util/observer/global-function-observer-example.ts
@@ -1,6 +1,6 @@
 import dotenv from "dotenv";
 import {
-  OpenAITextGenerationModel,
+  OpenAICompletionModel,
   generateText,
   setGlobalFunctionObservers,
 } from "modelfusion";
@@ -13,7 +13,7 @@ async function main() {
   setGlobalFunctionObservers([customObserver]);
 
   const text = await generateText(
-    new OpenAITextGenerationModel({
+    new OpenAICompletionModel({
       model: "gpt-3.5-turbo-instruct",
       maxCompletionTokens: 50,
     }),