+++ test-environments/esm/src/index.js
@@ -1,4 +1,4 @@
-import { OpenAITextGenerationModel, generateText } from "modelfusion";
+import { OpenAICompletionModel, generateText } from "modelfusion";
 import dotenv from "dotenv";
 
 dotenv.config();
@@ -6,7 +6,7 @@ dotenv.config();
 const OPENAI_API_KEY = process.env.OPENAI_API_KEY ?? "";
 
 (async () => {
-  const textGenerationModel = new OpenAITextGenerationModel({
+  const textGenerationModel = new OpenAICompletionModel({
     apiKey: OPENAI_API_KEY,
     model: "gpt-3.5-turbo-instruct",
     settings: { temperature: 0.7 },